authoerizing with gate on(evnt,attendee)

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -12,7 +12,13 @@ use Illuminate\Http\Request;
 class AttendeeController extends Controller
 {
  use CanLoadRelationship;
- protected array $relations =["user"];
+    protected array $relations =["user"];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except('index','show');
+    }
+
     public function index(Event $event)
     {
         $attendees = $this->LoadRelationship($event->attendees()->latest());
@@ -55,6 +61,7 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event,Attendee $attendee)
     {
+        $this->authorize('attendee-delete',[ $event , $attendee]);
         $attendee->delete();
 
         return response(status:204);
